requirements.txt
python-binance==1.0.19
pandas
requests
python-dotenv
